Adding support for additional plot types. Now all variable types can be plotted rather than just scalars as pseudocolor plots. A Combo near the top of the VisIt Plot Viewer allows the user to set the plot type.
Removing the commented out code for the previous, play/pause, and next
actions that cycled through the elements in the list in the VisIt Plot
Viewer. This code was previously commented out in favor of the
TimeSliderWidget. This also required bit of minor refactoring.

Adding a missing copyright notice to the MOOSETreeCompositeView.

Signed-off-by: Taylor Patterson <pattersontc@ornl.gov>