Adding code to fix a UI bug in the ICESectionPage.
I cleaned all of the code that creates the DataComponent,
TableComponent, and MatrixComponent sections. There was a bit of
unnecessary code there.

Added missing documentation and removed a lot of unnecessary tags and
empty lines.

ScrollClientComposite is a custom Composite that will re-adjust its size
based on the width of a parent SharedScrollComposite (which is used in
the Eclipse ScrolledForm) or a minimum size. Effectively, the default
preferred size of the DataComponent Sections in the ICESectionPage are
ignored because sometimes their preferred sizes are absurdly long (e.g.,
a Combo with a very long String).

PadFinder is used to find horizontal padding resulting from FillLayouts
and GridLayouts lying between the ScrollClientComposite and its ancestor
SharedScrolledComposite. This makes the resizing in
ScrollClientComposite a little smoother and more accurate.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>