Fixed a bug with the ICEFormEditor. Previously, it would pass the first available TreeComposite [it got from the Form] to the TreeCompositeViewer.
With recent changes to the MOOSEModel, there were more than one
TreeComposite, and the incorrect one was being passed to the Tree
Viewer.

To resolve this, I added a new method to ICEFormEditor that provides the
ID of the desired TreeComposite (usually retrieved from a model).
MOOSEFormEditor overrides this method to look for the correct MOOSE data
tree. The ICEFormEditor finds the corresponding TreeComposite, or the
first available TreeComposite if the desired one cannot be found, and
sends it to the TreeCompositeViewer.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>