Simplified the MOOSEFormEditor's interaction with its PlantBlockManager, which maps the "Components" block to a set of PlantComponents.
There was a lot of unnecessary work done that let the MOOSEFormEditor
listen to the MOOSE app selection in the MOOSETreeCompositeView.

Moved the code that creates the plant view to a couple of private
methods instead of an anonymous class inside addPlantPage(). On that
note, all of the additions to the plant view's ToolBar are done in the
private method fillPlantViewToolBar(ToolBarManager).

This fixes bug 455367 where the plant view renders empty until
components are unchecked and then checked again.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>