Renamed the native SFR packages and folders to be consistent with the other native packages. (I started on this because the native and non-native SFR folders had the same name, which confused SDC's git repo tasks)
Fixed all of the CMakeLists that were referring to "license" rather than
"LICENSE", which is in the main ICE directory. This was breaking the
CMake build.

Commented out the Doxygen bits from the main CMakeLists because the
UseDoxygen script and its input file are not in the git repo.

Fixed a memory leak in the native SFR Ring class. Essentially there were
a couple of undefined primitives when certain constructors were used.

With these changes, I was able to build ICE IO and the SFR library. All
tests pass.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>