Added a flag to the Visualization Service preference page to enable automatically connecting to default connections.
Added an AbstractVizService to include a getter for the preference
store.

Added a ConnectionAdapter class to manage interactions with connection
APIs (e.g., VisItSwtConnection and friends). The VisItVizService has its
own implementation of the adapter. The main purpose of the adapter is to
make the viz service code much easier to read. Have the adapter
separated will also be useful when the service allows client code to
have multiple viz connections to different hosts simultaneously.

Hooked up the table of connections to the VisItVizService preference
page. The VisItVizService now uses the default connection selected from
that table. If the default connection is changed, the associated
VisItPlots will attempt to reconnect. (I tested this out, and it worked
when switching the connection from remote to local!)

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>