Was able to finally fix the widget disposed error I was getting when trying to plot files.
When a ICEResourcePage was closed and it disposed of its widgets, the
page wasn't removed from the workbench's SelectionService listeners. So
everytime the SelectionService updated its listeners at a later point
(like when another MOOSE Launcher tried to plot files), it also
attempted to update the old ICEResourcePages with disposed widgets,
causing the error.

I added a DisposeListener onto the ICEResourcePage's main composite, so
when it's the composite's time to go, the ICEResourcePage is also
unregistered to get updates from the SelectionService.

Fixes bug 459707.

Signed-off-by: Anna Wojtowicz <wojtowicza@ornl.gov>