Pushed a commit that addresses bugs 460033, 460034, and 460035. All of these were cause by the viz service attempting to connect to an invalid default connection.
Now, if the default connection is not configured, the VisItPlot updates
itself accordingly and the viz service does not attempt to connect to
it. 

I also added code to the VisItPreferencePage to notify the viz service
of removed, changed, and added connections. Although it turns out this
code is not necessary to fix the aforementioned bugs, it will come in
handy when we have multiple connections open simultaneously.

Signed-off-by: jdeyton <jordan.deyton@gmail.com>