Fixed two bugs in the MOOSE Tree View that caused unnecessary updates to the MOOSE Model.
The first unnecessary update could occur if the active MOOSE Model
Builder is deactivated then activated by clicking on other views in the
workbench. I added a check to the IPartListener2 to only update if the
activated MOOSE Model Builder is a new one.

The second unnecessary update could occur if the same value was selected
from the MOOSE app Combo in the MOOSE Tree View. This was fixed by
adding a little logic to do nothing when the  Combo's value does not
actually change.

Signed-off-by: jdeyton <jordan.deyton@gmail.com>