Rewrote parts of the ICEResourcePage to handle drawing plots only when possible. Plots are created/loaded when the ResourceComponent is set or updated, but drawn only when selected the first time. After being drawn once, a plot doesn't need to be drawn again.
Changed the plots to not "tile". The selected resource in the Resource
View will be rendered alone. We can revisit tiling later after we come
up with a good way to be able to remove them (or put them into separate
Plot Editors).

From what I can tell, when a plot is invalid, it is caught either at
create time or at draw time. If a plot cannot be created or drawn, the
resource is sent to the browser.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>