Fixed some annoying bugs with the ParaView widget: (1) Selecting a different plot type will unplot the previously selected plot type. After I add checkboxes to the ActionTree, I'll revisit this and allow multiple plots to be selected. (2) Changing the plot types would produce multiple legends. I added code to hide the additional legends and lay them out automatically. (3) Added an extra render after the widget is painted because sometimes the ParaView widget thinks it is drawing in a very small (as in a few square pixels) container, and returns a very low quality image.
I also--at least temporarily--removed the ability to select the mesh for
the file. That is because ParaView allows you to select multiple meshes,
but you must have at least one mesh selected at all times. I haven't
quite figured out how to expose this functionality in the UI, and it's a
bit different from VisIt's approach.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>