Added a "None" option to the list of MOOSE Model executables to force the user to select one. Fixes bug 464646.
The "file" parameter on the Mesh block will correctly comment itself out
if the block type is anything other than FileMesh (and vice versa).
Also, if a "file" parameter is ever manually added at any point, it will
be registered with the MOOSE Model to send updates. Fixes bug 459693.

Also added to the AdaptiveTreeComposite.setType(...) method to correctly
re-set the active data node after the type is changed.

Signed-off-by: Anna Wojtowicz <wojtowicza@ornl.gov>