Changed DrawnPlot to a Composite. It now handles its own layout.
Updated the methods that remove the drawn plots to go through a single
method where possible. It uses a list (so PlotGridComposite clients can
remove by insertion order) and a map (for fast removal when the
DrawnPlot is known but not its index, primarily for internal use or when
it is disposed).

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>