Committing NewKernelHandler. This class is invoked when the user right clicks on their forked MOOSE application project and selects New > Moose Object > Kernel. This prompts them for a new Kernel name, and produces a header and source file stubbed out for that new Kernel subclass.
Updated documentation throughout this and the ForkStorkHandler.
Reorganized the widgets.moose plugin to have a actions and wizards
package. Moved the YAML Gen and Fork stuff to the wizards package, and
the NewKernelHandler to the actions package. In the future we can
populate this actions package with a bunch of new Moose object
generators. 

Signed-off-by: Alex McCaskey <mccaskeyaj@ornl.gov>