Fixing new Generate YAML action bug.
The JobLauncher now throws an error if any of the file entries are empty
or null. This was put in before checking that the generate yaml action
worked, which it doesn't because it uses no files. 

Since this action does not need any files, I'm just clearing them from
the file DataComponent in the MooseLauncher Form. 

Signed-off-by: Alex McCaskey <mccaskeyaj@ornl.gov>