Added a protected method to ParaViewPlot so the rendering code can access the IParaViewProxy.
Added a getter to IParaViewProxy so code that uses it may access its
underlying view ID.

Started working on integrating the IParaViewProxy into the
ParaViewPlotRender. The work is happening in a ParaViewProxyRender,
which will eventually replace ParaViewPlotRender.

Overrode setFeatureImpl(...) in ExodusProxy so that the proper plot
category can be set.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>