Added a play/pause button and tests for it.
It has not yet been decorated, so it is a regular button with text on
it.
It will halt playback either when the play/pause button is pressed or
when another time widget is selected.
It will currently loop back to the first timestep, and it plays in 1
second increments.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>