Added a fix to address a zoom bug.
Refresh events on the associated ParaViewCanvas keep getting triggered
after the mouse has been zoomed. I added a listener to notify the client
that the view is no longer zooming when the focus is lost.

I also moved the code in the mouse listeners to separate instances of
mouse listeners. ParaViewMouseAdapter no longer implements the SWT mouse
listener interfaces, so it can no longer be erroneously added to other
controls as listeners.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>