Updating the ParaViewVizPreferencePage to allow multiple connections.
The "default connection" property has been removed from the preference
page. Updates to the table in the preference page now propagate to the
VizConnectionManager in the ParaViewVizService automatically without the
use of static variables. In other words, the ParaViewVizService now
stores and can use multiple connections stored in the Eclipse
preferences.

Also removing an unnecessary comment block from VizConnectionManager.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>