Replaced the existing VisIt preferences page.
Had to update VizConnectionPreferencePage to create the new
VisItConnectionPreferencePage.
Deleted the old VisItConnectionTable and VisItPreferencesPage.
Fixed a bug in the VizConnectionPreferencePage - it said "ParaView
Connection Preferences". Now both the ParaView and VisIt preference
pages have a custom description.
Had to disable a lot of code in VisItVizService and VisItPlot to compile
after recent changes.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>