Updated Connection-based viz service classes to better handle series.
Added IPlotListener, which can be registered for update events against
an AbstractPlot (this is used to notify UI components when the CSVPlot
and VisItPlots load their data). This also replaced the existing
ICSVPlotListener.

Streamlined the connection-based classes to follow the recent changes to
PlotComposite. This includes changing ConnectionPlotRender to
ConnectionPlotComposite and working on updating ConnectionPlot.

Minor method rearrangement and document updates in PlotComposite.
Deprecated old classes until their sub-classes are removed from the
ParaView bundle.