Updated the VisItVizService to use the new ISeries infrastructure.
This also includes updates to use the updated connection widgets
(ConnectionPlot, ConnectionPlotComposite).

The VisItPlot now reloads its data when its connection or data source
are reset. It also appears to work well when, for instance, an editor
with a mesh file is open at startup. You can see the PlotEditor saying
it's waiting on the connection to be established, then it waits on the
data to be loaded, then it can finally draw the mesh.