Prepared code for re-use in solution to bug 474744.
Since the LWR HDF IO code is being updated to read like the
SFReactorIOHandler, there's a lot of potential for code re-use. Much of
the HDF IO code in the SFReactorIOHandler was already copied to
HdfIOFactory, so I made SFReactorIOHandler a sub-class of it to re-use
the code.

Also fixed a bug regarding reading strings from HDF. HDF uses native,
null-terminated strings. The LWR writer currently does not write
null-terminated strings, so I added logic to HdfIOFactory to check for a
null-string when reading in string attributes.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>