Fixed a bug in LWRDataProvider.
If there are multiple features in an LWRDataProvider, equality checks
fail when writing to then reading from HDF, even though the data content
is the same. I changed the equals method in LWRDataProvider to try to
find a corresponding FeatureSet in the other LWRDataProvider so that the
order in which features are added has no effect on equality as long as
the same features are added.

Signed-off-by: Jordan Deyton <deytonjh@ornl.gov>