Bug fixes for MOOSE-App URI Value and MultiLauncher NPE
Previously, for remote applications the EntryComposite set the app URI
to the returned IFileStore's URI.toString - which was something like
ssh://CONNECTION_NAME/PATH/TO/APP where connection name was just the
name of the connection and not the hostname, which was needed to figure
out which IRemoteConnection to use in the launch. This commit fixes this
and sets the entry value to ssh://hostname/path. 

This commit also changes the MOOSE Workflow ItemType to Application,
which fixes the MultiLauncher NPE in bug 478177. 

Cleaned up CheckMooseInputAction a bit.

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>