Fixed major Real-time CSV updating bug, added other improvements
Fixed a bug in the CSV Viz stuff that caused real-time plot updating for
the MOOSE item to break. 

Fixed bug where 2 CSV plots would open for the MOOSE item postprocessors
if the launch failed the first time. 

Set up the double click action on the ICEResourceView to also open the
ICEResourcePage instead of just creating the ICEResource Plot and not
displaying it. 

Fixed a bug in the CheckMooseInputAction related to searching for
errors. 

Made the MOOSE input file write a little less verbose. 

Setup the MOOSE update mechanism to use IResources instead of Java Files
and hooked up the ICEResourcePage to be a IResourceChangeListener to get
notified with VizResources are changed in the Workspace. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>