Added SWTBot tests and made minor UI improvements
This commit adds SWTBot tests for the Materials Database, MOOSE
Workflow, and Reflecftivity Model. They are not currently part of the
build and need a specialized pom file activating the services required
by an ICE workbench before they could be added to it. It also addresses
several minor UI bugs with those three items:

Fixed an issue with the Materials Database accepting non-positive
numbers as a material's amount.

Fixed several issues with the MOOSE Workflow not properly handling its
error messages and fields' dirty states.

Fixed the Reflectivity Model so that collapsing and expanding the
NatTable no longer increases the widgets' size. 

Fixed an issue where the Reflectivity Model's NatTable graphical
representation could be stale because it was not being refreshed after
each operation on it.

Set the Reflectivity Model to automatically open the Properties view
with itself in focus. Previously, it had set focus to the Resources
view, leaving the Properties view empty until the user clicked on the
Reflectivity Model. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>