Added new extensible Developer Menu
This commit adds a new extension point that lets users easily add new
developer-related actions to the ICE Developer Menu. 

Added an extension for ICE and MOOSE - with MOOSE providing the existing
developer actions we had before. I've turned off the action buttons that
were in the toolbar for MOOSE previously. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>