Fixed default project bug and reworked the XMLPersistenceProvider.
IPersistenceProviders, and thus the XMLPersistenceProvider, are no
longer required to create their own projects. It is assumed that they
will require an IProject - although that is not enforced - and that is
provided to them after they are started. I also fixed the construction
of the default Item storage project in the Core. It was putting all the
Items in the default project instead of the itemDB project.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>