Updated Handler Hierarchy and Added Build ICE Command
This commit updates the developer abstract handler hierarchy with a
GitCloneHandler that simple clones a repository, but provides a hook for
subclasses to provide PostCloneTasks, such as importing projects or
creating make targets. 

This commit also provides a new Build ICE command that grabs the
ILaunchManager and uses it to launch ICE's Maven Build launch
configuration files. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>