Set up clone ICETests dev menu action.
This commit fixes a bug where the cloned repo wasn't added to the git
repositories view. Also, it adds code to check for the user credentials
on a git repo if need be. 

Added an extension point that let's users clone the ICETests repo from
github. To acheive this, I added the ability to provide custom
parameters to a given command. The ICETests extension point uses this to
specify the clone directory as @user.home.

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>