Fixed client and core bugs due to Declarative Services issues.
Mic drop! Got the Core and Client working smoothly together in such a
way that the ItemViewer now works again, the workspace dialog still
works, the HttpService still works, neither are set to auto-start in the
run configuration and both are registered as services. The primary
problem was in the way that Declarative Services are activated by the
framework.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>