Fixing bugs encountered when using MOOSE Item
First bug - IReader used by MOOSEModel was null due to the fact that we
are switching to extension points for our IReader/IWriter interface
realizations. Set up the IOService to add all available
IReaders/IWriters from extension points to itself

Second bug - Found that when an Item XML file was in the project, and I
went to refresh or close the workspace it would hang and throw up a
dialog that said waiting on background task forever. Turns out the
XMLFormContentDescriber was stuck in an infinite loop reading the XML
contents to tell if it was a ICE Item or not. Fixed that and everything
works as expected. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>