Fixed Remote YAML and ICE Updater bug
Updated the RemoteYamlSyntaxGenerator to be more robust. Set it up to
create a new shell script to execute and generate the yaml/syntax files. 

Fixed the bug where real time plots failed to appear on form editor.
Somehow the javax.ws.rs 2.0.1 bundle got turned on in the launch config
that caused a clash with the 1.1.1 bundle. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>