Starting work on major ICEFormEditor bug.
So the ICEFormEditor.init method gets called when you double click a ICE
XML file, and the Client loads the Form for the editor to use. But
nothing happens to make sure the Client is updated for form changes,
process, and cancel events. Starting work on getting the MOOSE Item to
successfully load from file. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>