Committing fix for bug 481958 javax.ws.rs version error.
The last commit added platform specific plugins to our target definition
and failed to resolve on Mac and Windows. I've reverted the target to
the previous configuration in this commit. I added a filter to the pom
platform resolution to just use the javax.ws.rs 1.1.1 plugin, not 2.0.1,
which revealed the 2.0.1 dependency came from the docker feature in the
target. I removed the docker tooling, and the 2.0.1 addition to the
target is gone. 

I'm leaving the filter for now to catch this bug in the future. We can
remove it though since no plugins are pulling in the 2.0.1 plugin. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>