Fixing bug 481961
This commit fixes the bug found in bug 481961 where an opened Item XML
File failed to execute processes due the lack of an IFormWidget object. 

This has been fixed with the addition of an addFormWidget method to the
IClient interface, and implementation on Client that adds it to the
formWidgetTable, and the ICEFormEditor setting a new
EclipseFormWidget(this) on its IClient reference if the input was a
FileEditorInput. 

Updated tests as well. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>