Refactored the new UI architecture to use a factory.
Updated the new UI code to use a factory, which could be - but currently
is not - obtained from an extension point. This required updates to the
ICEFormEditor to use the factory and to migrate the old construction
code out to the new factory. I also updated the page providers to use a
slightly different signature. I added a base class for default page
providers so that they only provider their pages and don't duplicate
code.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>