Setup ICE to take a http post of type FILE_PLOTEDITOR.
This commits enables users to point ICE to a data file via an http post
and have ICE open it in the PlotEditor if it can. 

Also adding a simple script to do this post via the command line. Just
run ./openICEPlotEditor fileName port. We can make this more robust
later for remote hostname posts. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>