Finished JobLauncher/IActionFactory Refactor.
This commit updates the JobLauncher to use the IActionFactory reference
to perform job launches. It creates a list of Actions based on the data
in the action data map and then passes those Actions to a custom Eclipse
Job class for execution and monitoring. Actions now post text to the
Eclipse Console. Added/Modified RemoteFileUploadAction,
RemoteExecutionAction, RemoteFileDownloadAction, LocalExecutionAction,
and LocalFilesCopyAction. 

This commit also fixes bugs in MOOSEModel (non-unique keys in app
preferences) and MOOSELauncher (failed to reset host in hosts table to
localhost after remote execution).

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>