Added run configurations for SWTBot tests
Added individual run configurations for the SWTBot tests with the
correct set of plug-in dependencies. These are for Windows only, as I'm
intending to create one bulk run configuration for each OS after dealing
with issues that are causing them to fail when all are run from a single
configuration file.

Also made minor changes to the tests to make them compatible with the
latest changes.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>