Fixed repeated pipe bug
Work around for bug where pipes are being read from the composite
repeatedly with different locations. The converter now checks if a pipe
with that id already exists in the plant before placing it in the tree.
Since the first one read was always the correct one, this essentially
fixes the bug.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>