Reworked the build. Fixed many Maven and Tycho sins.
ICE now uses an aggregator pom instead of combining the list of modules
with the parent. Both the parent and the aggregrator are in their own
bundles. Every other bundle has been redirected to point to the new
parent and all ICE bundles are now Maven projects. It is no longer
necessary to install the build bundle. Now, just go into the
org.eclipse.ice.aggregator directory and run "mvn clean verify" from the
CLI or run the build with these goals from the Maven run configuration.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>