Started DockerExecutionAction
DockerExecutionAction is a subclass of RemoteExecutionAction that does
some preprocessing before a remote execution - this preprocessing
creates a detached docker container with port 22 mapped to a local port
on the machine, then invokes the usual remote execution action with the
DOCKER_HOST as the hostname and the local port instead of 22.

Note this is not fully functional yet, still working on it.

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>