Plot Editor enhancement
The CSV Plot Editor now can save edits to its data and will redraw
itself based on the new file. The code for creating the pages for the
plot is now in the IPlot instead of the IVizService.

Also fixed a problem in the PlotEditorTester where the folder
CSVVizService wouldn't open if the test took too long, as the name would
change to "> CSVVizService [ICETests master]"

Signed-off-by: Robert Smith <SmithRW@ornl.gov>