Fixed launch button with e4 extensions. Added more spies to target.
The reflectivity launcher now posts its button to the toolbar using E4
instead of 3.x. This seemed like a better way to handle dependency
injection to get the client OSGi service. Note that getting the toolbar
button requires adding a trimContribution and getting it to stay in one
place requires setting position to after=additions.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>