Fixing bug #166 where FileEntry would not update when file imported.
Setting up the FileEntry to be an IResourceChangeListeneer that updates
its allowed values when a file is added/changed in the project space.
This only works if clients invoke FileEntry.setProject with their Item's
current IProject instance. Otherwise, resources updates are ignored. 

Added test to FileEntryTester and verified it worked with
ReflectivityModel wave_vector file. 

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>