Closed #180.
I refactored the ItemProcessor so that it now implements an
IItemProcessor interface. The Client now takes an IItemProcessor in the
test constructor, and the ClientTester provides a FakeItemProcessor. The
FakeItemProcessor runs most (but not all) of the code of the
ItemProcessor on the test thread, thus fixing #180. I tested this
thoroughly on my machine, both with the test framework and the product.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>