Fixed RemoteAction bug for using old connections
Fixed a bug where RemoteAction was trying to find pre-existing
connections by looking up the address for the current connection's name
instead of its hostname. It now simply compares the two connections'
names. This allows a user to, for example, configure two different
connections to the same machine but which use different ports.

Signed-off-by: Robert Smith <smithrw@ornl.gov>