Address #348.
This bug removes all the old viz service factory code from the MOOSE and
Reflectivity UI bundles. I have not noticed any problems because of this
yet, but since EAVP is busted because of the DS changes there may be
"unknown unknowns." Or "known unknowns." I'm not sure which but I just
Rumsfelded this commit message, whoo!

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>