Finished ICEModelFactory basic implementation.
Made a basic implementation of ICEModelFactory, which is the new name
for ModelFactory. I made the name change so that it would not collide
with Apache Jena's ModelFactory. Also committed a basic test for it that
checks for null values returned by the factory methods. Finally, I
removed the unused AppTest class.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>