Replace field alias concept with simpler system
Aliases were introduced to address the problem presented by privateId's
getter being getUUID() instead of getPrivateId(). The concept of Aliases
was needlessly complicated to solve this problem and otherwise unused
anywhere else in the project. In it's place, "var" was introduced on
Field that can hold the variable's name when it differs from the name of
the field itself, i.e. in the privateId scenario, the "var" is
"privateId" and the field name is "UUID"

Signed-off-by: Daniel Bluhm <bluhmdj@ornl.gov>